[io] Explode io.rs into a bunch of files
io.rs was getting a bit too kitchen-sinky, making it diffucult
to find a good place to put new functions. This change breaks it up into
six files, keeping the most common names readily accessible at
manticore::io.
